Compress classes.

+++ codingame/contests/2105/action.h
@@ -12,7 +12,7 @@ class Action {
   ActionType type;
   uint8_t value1, value2;
 
-  Action(ActionType _type = WAIT, unsigned _value1 = 0, unsigned _value2 = 0)
+  Action(ActionType _type = WAIT, uint8_t _value1 = 0, uint8_t _value2 = 0)
       : type(_type), value1(_value1), value2(_value2) {}
 
   void Write() const {
