+++ codingame/contests/2105/cell.h
@@ -8,18 +8,18 @@
 
 class Cell {
  public:
-  unsigned index;     // 0 is the center cell, the next cells spiral outwards
-  unsigned richness;  // 0 if the cell is unusable, 1-3 for usable cells
-  std::vector<unsigned> neighbors;  // the index of the neighbouring cell
+  uint8_t index;     // 0 is the center cell, the next cells spiral outwards
+  uint8_t richness;  // 0 if the cell is unusable, 1-3 for usable cells
+  std::vector<uint8_t> neighbors;  // the index of the neighbouring cell
 
   void Read() {
     std::cin >> index >> richness;
-    neighbors = nvector::Read<unsigned>(6);
+    neighbors = nvector::Read<uint8_t>(6);
     std::cin.ignore();
   }
 
   static std::vector<Cell> ReadCells() {
-    unsigned total;
+    uint8_t total;
     std::cin >> total;
     std::cin.ignore();
     assert(total == 37);
